separated code into layers according to onion architecture
