+++ .sequelizerc
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  'config': path.resolve('src/config/index.js'),
-  'models-path': path.resolve('src/db/models'),
-  'migrations-path': path.resolve('src/db/migrations'),
+  'config': path.resolve('src/common/config/index.js'),
+  'models-path': path.resolve('src/infrastructure/db/models'),
+  'migrations-path': path.resolve('src/infrastructure/db/migrations'),
 };
