+++ src/common/utils/emailTemplates.js
@@ -1,4 +1,4 @@
-const config = require('../config/index');
+const config = require('../config');
 
 function buildConfirmEmail(token)
 {
