+++ src/common/utils/strings.js
