+++ src/infrastructure/providers/weather-providers/weatherApiProvider.js
@@ -1,8 +1,8 @@
-const IWeatherProvider = require('../../domain/interfaces/providers/weatherProviderInterface');
-const config = require('../../config/index');
+const IWeatherProvider = require('../../../domain/interfaces/providers/weatherProviderInterface');
+const config = require('../../../common/config');
 const WEATHER_API_KEY = config.weatherApiKey;
 
-const Result = require('../../domain/types/result');
+const Result = require('../../../domain/types/result');
 
 class WeatherApiProvider extends IWeatherProvider
 {
