+++ src/infrastructure/providers/weather-providers/weatherProviderManager.js
@@ -1,5 +1,5 @@
-const IWeatherProvider = require('../../domain/interfaces/providers/weatherProviderInterface');
-const Result = require('../../domain/types/result');
+const IWeatherProvider = require('../../../domain/interfaces/providers/weatherProviderInterface');
+const Result = require('../../../domain/types/result');
 
 class WeatherProviderManager extends IWeatherProvider
 {
