+++ src/presentation/routes/weather.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { weatherApiController } = require('../setup');
+const { weatherApiController } = require('../../setup');
 
 // expose the weather api endpoint
 router.get('/weather', weatherApiController.getWeather);
