+++ test/integration/weather.test.js
@@ -2,11 +2,11 @@ const { expect } = require('chai');
 const request = require('supertest');
 const express = require('express');
 
-const WeatherApiController = require('../../src/controllers/weatherApiController');
+const WeatherApiController = require('../../src/presentation/controllers/weatherApiController');
 const WeatherServiceWithCacheAndMetrics = require('../../src/services/weatherService');
 const GetWeatherUseCase = require('../../src/use-cases/weather/getWeatherUseCase');
 const MockWeatherProviderManager = require('../mocks/providers/weatherProviderManager.mock');
-const { redisClient, connectToRedisWithRetry } = require('../../src/utils/redisClient');
+const { redisClient, connectToRedisWithRetry } = require('../../src/common/utils/redisClient');
 const SimpleCounter = require('../mocks/utils/metrics.mock');
 
 connectToRedisWithRetry();
